# shared/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FastShip",
    "version": "0.1.0"
  },
  "paths": {
...
    "/ops/heatmap": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Heatmap",
        "operationId": "heatmap_ops_heatmap_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/HeatmapCell"
                  },
                  "title": "Response Heatmap Ops Heatmap Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ops/complaints": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Complaints",
        "operationId": "complaints_ops_complaints_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ComplaintOut"
                  },
                  "title": "Response Complaints Ops Complaints Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ops/orders/{order_id}/reassign": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "Reassign",
        "operationId": "reassign_ops_orders__order_id__reassign_post",
        "parameters": [
          {
            "name": "order_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Order Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ops/payments/{payment_id}/resolve": {
      "post": {
        "tags": [
          "ops"
        ],
        "summary": "Resolve Payment",
        "operationId": "resolve_payment_ops_payments__payment_id__resolve_post",
        "parameters": [
          {
            "name": "payment_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Payment Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolvePaymentRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/ops/reports/summary": {
      "get": {
        "tags": [
          "ops"
        ],
        "summary": "Reports Summary",
        "operationId": "reports_summary_ops_reports_summary_get",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SummaryReport"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ComplaintOut": {
        "properties": {
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "payment_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payment Id"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "order_id",
          "payment_id",
          "reason",
          "status"
        ],
        "title": "ComplaintOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "HeatmapCell": {
        "properties": {
          "lat_bucket": {
            "type": "number",
            "title": "Lat Bucket"
          },
          "lng_bucket": {
            "type": "number",
            "title": "Lng Bucket"
          },
          "shipper_count": {
            "type": "integer",
            "title": "Shipper Count"
          }
        },
        "type": "object",
        "required": [
          "lat_bucket",
          "lng_bucket",
          "shipper_count"
        ],
        "title": "HeatmapCell"
      },
...
      "PaymentMethod": {
        "type": "string",
        "enum": [
          "cod",
          "wallet",
          "card"
        ],
        "title": "PaymentMethod"
      },
      "PaymentOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "order_id": {
            "type": "string",
            "format": "uuid",
            "title": "Order Id"
          },
          "method": {
            "$ref": "#/components/schemas/PaymentMethod"
          },
          "status": {
            "$ref": "#/components/schemas/PaymentStatus"
          },
          "amount": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Amount"
          },
          "held_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Held At"
          },
          "release_due_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Due At"
          },
          "released_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Released At"
          },
          "refunded_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Refunded At"
          },
          "dispute_reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Dispute Reason"
          }
        },
        "type": "object",
        "required": [
          "id",
          "order_id",
          "method",
          "status",
          "amount",
          "held_at",
          "release_due_at",
          "released_at",
          "refunded_at",
          "dispute_reason"
        ],
        "title": "PaymentOut"
      },
      "PaymentStatus": {
        "type": "string",
        "enum": [
          "pending",
          "held",
          "released_to_merchant",
          "refunded",
          "disputed"
        ],
        "title": "PaymentStatus"
      },
...
      "ResolvePaymentRequest": {
        "properties": {
          "release": {
            "type": "boolean",
            "title": "Release"
          }
        },
        "type": "object",
        "required": [
          "release"
        ],
        "title": "ResolvePaymentRequest"
      },
...
      "SummaryReport": {
        "properties": {
          "total_orders": {
            "type": "integer",
            "title": "Total Orders"
          },
          "completed_orders": {
            "type": "integer",
            "title": "Completed Orders"
          },
          "failed_orders": {
            "type": "integer",
            "title": "Failed Orders"
          },
          "active_orders": {
            "type": "integer",
            "title": "Active Orders"
          },
          "total_revenue": {
            "type": "string",
            "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$",
            "title": "Total Revenue"
          },
          "disputed_payments": {
            "type": "integer",
            "title": "Disputed Payments"
          }
        },
        "type": "object",
        "required": [
          "total_orders",
          "completed_orders",
          "failed_orders",
          "active_orders",
          "total_revenue",
          "disputed_payments"
        ],
        "title": "SummaryReport"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}